pyproject.toml
[tool.poetry]
name = "jyyslide-md"
version = "0.1.0"
description = ""
authors = ["zweix123 <1979803044@qq.com>"]
readme = "README.md"
packages = [{include = "jyyslide_md"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
jinja2 = "^3.1.2"
markdown = "^3.4.1"
pyquery = "^2.0.0"
pygments = "^2.14.0"
lxml = "^4.9.4"
pyyaml = "^6.0.1"
types-markdown = "^3.4.2.6"
types-requests = "^2.28.11.17"
types-pyyaml = "^6.0.12.12"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
